[refactor] api 패키지의 controller 패키지 삭제
