+++ src/main/java/junseok/snr/wallet/api/dto/CreateWalletResponseDto.java
@@ -1,4 +1,4 @@
-package junseok.snr.wallet.api.controller.dto;
+package junseok.snr.wallet.api.dto;
 
 import junseok.snr.wallet.domain.WalletType;
 import lombok.AllArgsConstructor;
